# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgroGestion API",
    "description": "API para gestionar cultivos de principio a fin: siembras, ciclos, conteo de plantas, vivero, riesgos y eventos adversos. Todos los errores usan el mismo formato: decida siempre con el campo `error`, nunca con `message`.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/consultas": {
      "post": {
        "tags": [
          "10. Asistente"
        ],
        "summary": "Preguntarle al asistente",
        "operationId": "preguntar_api_v1_consultas_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaCrear"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultaSalida"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "429": {
            "description": "Llegó al máximo de consultas del día.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "10. Asistente"
        ],
        "summary": "Historial de consultas",
        "operationId": "historial_api_v1_consultas_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "description": "Texto en la pregunta",
              "title": "Q"
            },
            "description": "Texto en la pregunta"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Registros a saltar",
              "default": 0,
              "title": "Skip"
            },
            "description": "Registros a saltar"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Tamaño de página, máximo 50",
              "default": 10,
              "title": "Limit"
            },
            "description": "Tamaño de página, máximo 50"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ConsultaResumen_"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/consultas/{consulta_id}": {
      "get": {
        "tags": [
          "10. Asistente"
        ],
        "summary": "Una consulta y su respuesta",
        "operationId": "ver_api_v1_consultas__consulta_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "consulta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Consulta Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultaSalida"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/consultas/{consulta_id}/retroalimentacion": {
      "post": {
        "tags": [
          "10. Asistente"
        ],
        "summary": "Decir si la respuesta sirvió",
        "operationId": "valorar_api_v1_consultas__consulta_id__retroalimentacion_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "consulta_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Consulta Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetroalimentacionEntrada"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetroalimentacionSalida"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/asistente/calidad": {
      "get": {
        "tags": [
          "10. Asistente"
        ],
        "summary": "Calidad del asistente",
        "operationId": "calidad_api_v1_asistente_calidad_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalidadSalida"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o sesión vencida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "403": {
            "description": "El rol no tiene permiso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe o no pertenece al usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "409": {
            "description": "Choca con algo que ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Datos inválidos o regla de negocio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalidadSalida": {
        "properties": {
          "consultas": {
            "type": "integer",
            "title": "Consultas"
          },
          "con_valoracion": {
            "type": "integer",
            "title": "Con Valoracion"
          },
          "sirvieron": {
            "type": "integer",
            "title": "Sirvieron"
          },
          "no_sirvieron": {
            "type": "integer",
            "title": "No Sirvieron"
          },
          "equivocadas": {
            "type": "integer",
            "title": "Equivocadas"
          },
          "porcentaje_que_sirvio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Porcentaje Que Sirvio"
          },
          "aviso": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aviso"
          }
        },
        "type": "object",
        "required": [
          "consultas",
          "con_valoracion",
          "sirvieron",
          "no_sirvieron",
          "equivocadas",
          "porcentaje_que_sirvio"
        ],
        "title": "CalidadSalida"
      },
...
      "CausaProbable": {
        "properties": {
          "problema_id": {
            "type": "string",
            "format": "uuid",
            "title": "Problema Id"
          },
          "problema": {
            "type": "string",
            "title": "Problema"
          },
          "coincidencia": {
            "type": "string",
            "enum": [
              "alta",
              "media",
              "baja"
            ],
            "title": "Coincidencia"
          },
          "por_que": {
            "type": "string",
            "title": "Por Que"
          },
          "fuentes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fuentes"
          }
        },
        "type": "object",
        "required": [
          "problema_id",
          "problema",
          "coincidencia",
          "por_que",
          "fuentes"
        ],
        "title": "CausaProbable"
      },
...
      "ConsultaCrear": {
        "properties": {
          "siembra_id": {
            "type": "string",
            "format": "uuid",
            "title": "Siembra Id"
          },
          "texto": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 5,
            "title": "Texto"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "siembra_id",
          "texto"
        ],
        "title": "ConsultaCrear"
      },
      "ConsultaResumen": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "contexto_id": {
            "type": "string",
            "format": "uuid",
            "title": "Contexto Id"
          },
          "creado_en": {
            "type": "string",
            "format": "date-time",
            "title": "Creado En"
          },
          "pregunta": {
            "type": "string",
            "title": "Pregunta"
          },
          "valoracion": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "sirvio",
                  "no_sirvio",
                  "equivocado"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Valoracion"
          }
        },
        "type": "object",
        "required": [
          "id",
          "contexto_id",
          "creado_en",
          "pregunta"
        ],
        "title": "ConsultaResumen"
      },
      "ConsultaSalida": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "contexto": {
            "type": "string",
            "title": "Contexto"
          },
          "contexto_id": {
            "type": "string",
            "format": "uuid",
            "title": "Contexto Id"
          },
          "creado_en": {
            "type": "string",
            "format": "date-time",
            "title": "Creado En"
          },
          "pregunta": {
            "type": "string",
            "title": "Pregunta"
          },
          "respuesta": {
            "$ref": "#/components/schemas/RespuestaAsistente"
          },
          "mensaje_respuesta_id": {
            "type": "string",
            "format": "uuid",
            "title": "Mensaje Respuesta Id"
          },
          "valoracion": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "sirvio",
                  "no_sirvio",
                  "equivocado"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Valoracion"
          }
        },
        "type": "object",
        "required": [
          "id",
          "contexto",
          "contexto_id",
          "creado_en",
          "pregunta",
          "respuesta",
          "mensaje_respuesta_id"
        ],
        "title": "ConsultaSalida"
      },
...
      "DetalleError": {
        "properties": {
          "campo": {
            "type": "string",
            "title": "Campo"
          },
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          }
        },
        "type": "object",
        "required": [
          "campo",
          "mensaje",
          "tipo"
        ],
        "title": "DetalleError"
      },
...
      "ErrorRespuesta": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status_code": {
            "type": "integer",
            "title": "Status Code"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DetalleError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message",
          "status_code"
        ],
        "title": "ErrorRespuesta",
        "description": "Formato único de error: decida siempre con `error`, nunca con `message`."
      },
...
      "Page_ConsultaResumen_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConsultaResumen"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "has_more"
        ],
        "title": "Page[ConsultaResumen]"
      },
...
      "RespuestaAsistente": {
        "properties": {
          "lo_que_entendi": {
            "type": "string",
            "title": "Lo Que Entendi"
          },
          "causas_probables": {
            "items": {
              "$ref": "#/components/schemas/CausaProbable"
            },
            "type": "array",
            "title": "Causas Probables"
          },
          "que_hacer": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Que Hacer"
          },
          "tratamientos": {
            "items": {
              "$ref": "#/components/schemas/TratamientoValidado"
            },
            "type": "array",
            "title": "Tratamientos"
          },
          "mensaje_tratamiento": {
            "type": "string",
            "title": "Mensaje Tratamiento"
          },
          "cuando_llamar_al_tecnico": {
            "type": "string",
            "title": "Cuando Llamar Al Tecnico"
          },
          "aviso": {
            "type": "string",
            "title": "Aviso"
          },
          "hay_informacion": {
            "type": "boolean",
            "title": "Hay Informacion"
          }
        },
        "type": "object",
        "required": [
          "lo_que_entendi",
          "causas_probables",
          "que_hacer",
          "tratamientos",
          "mensaje_tratamiento",
          "cuando_llamar_al_tecnico",
          "aviso",
          "hay_informacion"
        ],
        "title": "RespuestaAsistente",
        "description": "Lo que el asistente entiende de la descripción y lo que hay validado para ese cultivo."
      },
      "RetroalimentacionEntrada": {
        "properties": {
          "valor": {
            "type": "string",
            "enum": [
              "sirvio",
              "no_sirvio",
              "equivocado"
            ],
            "title": "Valor"
          },
          "comentario": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comentario"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "valor"
        ],
        "title": "RetroalimentacionEntrada"
      },
      "RetroalimentacionSalida": {
        "properties": {
          "mensaje_id": {
            "type": "string",
            "format": "uuid",
            "title": "Mensaje Id"
          },
          "valor": {
            "type": "string",
            "title": "Valor"
          },
          "comentario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comentario"
          }
        },
        "type": "object",
        "required": [
          "mensaje_id",
          "valor",
          "comentario"
        ],
        "title": "RetroalimentacionSalida"
      },
...
      "TratamientoValidado": {
        "properties": {
          "problema": {
            "type": "string",
            "title": "Problema"
          },
          "producto_ica": {
            "type": "string",
            "title": "Producto Ica"
          },
          "descripcion": {
            "type": "string",
            "title": "Descripcion"
          },
          "fuente": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fuente"
          }
        },
        "type": "object",
        "required": [
          "problema",
          "producto_ica",
          "descripcion",
          "fuente"
        ],
        "title": "TratamientoValidado"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}